moved additional table functionality to db.go
